Separates supplement update form, logic and functions to another component. Reduces size of HostProfile component even further.
